Integrated governance pool functions (activate IP and mint tokens) into IPF and IP. Added tests for integrations.

+++ packages/governance-pool/contracts/interfaces/IGovernancePool.sol
@@ -21,7 +21,7 @@ interface IGovernancePool {
     function mintVotingTokens(
         address _investor,
         uint256 _amount,
-        uint256 _unlockTime
+        uint48 _unlockTime
     ) external;
 
     function unlockVotingTokens(address _investmentPool) external;
